+++ aframe/src/location-based/gps-projected-entity-place.js
@@ -1,6 +1,25 @@
-import * as AFRAME from 'aframe';
-
-AFRAME.registerComponent('gps-entity-place', {
+/** gps-projected-entity-place
+ *
+ * based on the original gps-entity-place, modified by nickw 02/04/20
+ *
+ * Rather than keeping track of position by calculating the distance of
+ * entities or the current location to the original location, this version
+ * makes use of the "Google" Spherical Mercactor projection, aka epsg:3857.
+ *
+ * The original location on startup (lat/lon) is projected into Spherical 
+ * Mercator and stored.
+ *
+ * When 'entity-places' are added, their Spherical Mercator coords are 
+ * calculated and converted into world coordinates, relative to the original
+ * position, using the Spherical Mercator projection calculation in
+ * gps-projected-camera.
+ *
+ * Spherical Mercator units are close to, but not exactly, metres, and are
+ * heavily distorted near the poles. Nonetheless they are a good approximation
+ * for many areas of the world and appear not to cause unacceptable distortions
+ * when used as the units for AR apps.
+ */
+AFRAME.registerComponent('gps-projected-entity-place', {
     _cameraGps: null,
     schema: {
         longitude: {
@@ -10,44 +29,50 @@ AFRAME.registerComponent('gps-entity-place', {
         latitude: {
             type: 'number',
             default: 0,
+        },
+        elevation: {
+            type: 'number',
+            default: 0
         }
     },
     remove: function() {
         // cleaning listeners when the entity is removed from the DOM
-        window.removeEventListener('gps-camera-origin-coord-set', this.coordSetListener);
         window.removeEventListener('gps-camera-update-position', this.updatePositionListener);
     },
     init: function() {
+        // Used now to get the GPS camera when it's been setup
         this.coordSetListener = () => {
             if (!this._cameraGps) {
-                var camera = document.querySelector('[gps-camera]');
-                if (!camera.components['gps-camera']) {
-                    console.error('gps-camera not initialized')
+                var camera = document.querySelector('[gps-projected-camera]');
+                if (!camera.components['gps-projected-camera']) {
+                    console.error('gps-projected-camera not initialized')
                     return;
                 }
-                this._cameraGps = camera.components['gps-camera'];
+                this._cameraGps = camera.components['gps-projected-camera'];
+                this._updatePosition();
             }
-            this._updatePosition();
         };
+        
 
+
+        // update position needs to worry about distance but nothing else?
         this.updatePositionListener = (ev) => {
             if (!this.data || !this._cameraGps) {
                 return;
             }
 
-            var dstCoords = {
-                longitude: this.data.longitude,
-                latitude: this.data.latitude,
-            };
+            var dstCoords = this.getAttribute('position');
 
             // it's actually a 'distance place', but we don't call it with last param, because we want to retrieve distance even if it's < minDistance property
-            var distanceForMsg = this._cameraGps.computeDistanceMeters(ev.detail.position, dstCoords);
+            // _computeDistanceMeters is now going to use the projected
+            var distanceForMsg = this._cameraGps.computeDistanceMeters(dstCoords);
 
             this.el.setAttribute('distance', distanceForMsg);
             this.el.setAttribute('distanceMsg', formatDistance(distanceForMsg));
+
             this.el.dispatchEvent(new CustomEvent('gps-entity-place-update-positon', { detail: { distance: distanceForMsg } }));
 
-            var actualDistance = this._cameraGps.computeDistanceMeters(ev.detail.position, dstCoords, true);
+            var actualDistance = this._cameraGps.computeDistanceMeters(dstCoords, true);
 
             if (actualDistance === Number.MAX_SAFE_INTEGER) {
                 this.hideForMinDistance(this.el, true);
@@ -56,6 +81,7 @@ AFRAME.registerComponent('gps-entity-place', {
             }
         };
 
+        // Retain as this event is fired when the GPS camera is set up
         window.addEventListener('gps-camera-origin-coord-set', this.coordSetListener);
         window.addEventListener('gps-camera-update-position', this.updatePositionListener);
 
@@ -78,38 +104,17 @@ AFRAME.registerComponent('gps-entity-place', {
      * Update place position
      * @returns {void}
      */
+
+    // set position to world coords using the lat/lon 
     _updatePosition: function() {
-        var position = { x: 0, y: this.el.getAttribute('position').y || 0, z: 0 }
-
-        // update position.x
-        var dstCoords = {
-            longitude: this.data.longitude,
-            latitude: this._cameraGps.originCoords.latitude,
-        };
-
-        position.x = this._cameraGps.computeDistanceMeters(this._cameraGps.originCoords, dstCoords);
-
-        this._positionXDebug = position.x;
-
-        position.x *= this.data.longitude > this._cameraGps.originCoords.longitude ? 1 : -1;
-
-        // update position.z
-        var dstCoords = {
-            longitude: this._cameraGps.originCoords.longitude,
-            latitude: this.data.latitude,
-        };
-
-        position.z = this._cameraGps.computeDistanceMeters(this._cameraGps.originCoords, dstCoords);
-
-        position.z *= this.data.latitude > this._cameraGps.originCoords.latitude ? -1 : 1;
-
-        if (position.y !== 0) {
-            var altitude = this._cameraGps.originCoords.altitude !== undefined ? this._cameraGps.originCoords.altitude : 0;
-            position.y = position.y - altitude;
-        }
-
+        var pos = this._cameraGps.latLonToWorld(this.data.latitude, this.data.longitude);
         // update element's position in 3D world
-        this.el.setAttribute('position', position);
+        //this.el.setAttribute('position', position);
+        this.el.setAttribute('position', {
+            x: pos[0],
+            y: this.data.elevation,
+            z: pos[1]
+        }); 
     },
 });